Add tests for retrieving the cart from the FarmerNPC
Created a test class for testing if the farmer will place a hand cart
when the player requests him to. The tests assert that the player must
be at least level 300 in order to retrieve the cart from Karl. This is
to enforce our requirement to allow only a subset of players to test
the feature.

These tests should fail at this commit.