Implement cart retrieval from the Farmer NPC
Added a new conversation trigger for the Farmer NPC so that he will
respond to the word cart. When a player says this, the NPC will check
the level of the player before giving them a cart. This is to enforce
the requirement that only a subset of players should be able to play
test this feature. We have currently set the barrier at level 300.
Relates to issue #16